Enable immediate submission after fe scheme change
Why
Prior to this changing the fee scheme required
hitting calculate once processed the form as
html and could be submitted as a JS request.

How
By not rerendring the entire form but just the components
of it we avoid cross-site-origin-request problems
which rails was handling by processing as html.